Remove non-elemental identifier types (#90)
