+++ src/Action/InputAction.php
@@ -4,12 +4,10 @@ namespace webignition\BasilModel\Action;
 
 use webignition\BasilModel\Identifier\IdentifierInterface;
 use webignition\BasilModel\Value\ValueInterface;
-use webignition\BasilModel\ValueContainerInterface;
-use webignition\BasilModel\ValueContainerTrait;
 
-class InputAction extends InteractionAction implements InputActionInterface, ValueContainerInterface
+class InputAction extends InteractionAction implements InputActionInterface
 {
-    use ValueContainerTrait;
+    private $value;
 
     public function __construct(
         string $actionString,
@@ -20,5 +18,10 @@ class InputAction extends InteractionAction implements InputActionInterface, Val
         parent::__construct($actionString, ActionTypes::SET, $identifier, $arguments);
 
         $this->value = $value;
+    }
+
+    public function getValue(): ?ValueInterface
+    {
+        return $this->value;
     }
 }
