+++ src/Action/InteractionActionInterface.php
@@ -2,8 +2,10 @@
 
 namespace webignition\BasilModel\Action;
 
-use webignition\BasilModel\IdentifierContainerInterface;
+use webignition\BasilModel\Identifier\IdentifierInterface;
 
-interface InteractionActionInterface extends ActionInterface, IdentifierContainerInterface
+interface InteractionActionInterface extends ActionInterface
 {
+    public function getIdentifier(): ?IdentifierInterface;
+    public function withIdentifier(IdentifierInterface $identifier): InteractionActionInterface;
 }
