+++ src/Identifier/IdentifierInterface.php
@@ -13,7 +13,5 @@ interface IdentifierInterface
     public function getParentIdentifier(): ?IdentifierInterface;
     public function withParentIdentifier(IdentifierInterface $identifier): IdentifierInterface;
     public function withName(string $name): IdentifierInterface;
-    public function isActionable(): bool;
-    public function isAssertable(): bool;
     public function __toString(): string;
 }
