--- src/ValueContainerInterface.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace webignition\BasilModel;
-
-use webignition\BasilModel\Value\ValueInterface;
-
-interface ValueContainerInterface
-{
-    public function getValue(): ?ValueInterface;
-}
