+++ tests/Identifier/IdentifierTest.php
@@ -7,9 +7,7 @@ use webignition\BasilModel\Identifier\Identifier;
 use webignition\BasilModel\Identifier\IdentifierInterface;
 use webignition\BasilModel\Identifier\IdentifierTypes;
 use webignition\BasilModel\Value\LiteralValue;
-use webignition\BasilModel\Value\ObjectValue;
 use webignition\BasilModel\Value\ValueInterface;
-use webignition\BasilModel\Value\ValueTypes;
 
 class IdentifierTest extends \PHPUnit\Framework\TestCase
 {
@@ -38,11 +36,6 @@ class IdentifierTest extends \PHPUnit\Framework\TestCase
                 'value' => new LiteralValue('.selector'),
                 'expectedPosition' => 3,
                 'position' => 3,
-            ],
-            'object value' => [
-                'type' => IdentifierTypes::PAGE_OBJECT_PARAMETER,
-                'value' => new ObjectValue(ValueTypes::PAGE_OBJECT_PROPERTY, '$page.url', 'page', 'url'),
-                'expectedPosition' => Identifier::DEFAULT_POSITION
             ],
         ];
     }
@@ -126,82 +119,6 @@ class IdentifierTest extends \PHPUnit\Framework\TestCase
                 'identifier' => new Identifier(IdentifierTypes::XPATH_EXPRESSION, $xpathExpressionIdentifierValue, 2),
                 'expectedString' => '"//foo":2',
             ],
-            'page model element reference, position null' => [
-                'identifier' => new Identifier(
-                    IdentifierTypes::PAGE_MODEL_ELEMENT_REFERENCE,
-                    new ObjectValue(
-                        ValueTypes::PAGE_ELEMENT_REFERENCE,
-                        'page_import_name.elements.element_name',
-                        'page_import_name',
-                        'element_name'
-                    )
-                ),
-                'expectedString' => 'page_import_name.elements.element_name',
-            ],
-            'page model element reference, position 1' => [
-                'identifier' => new Identifier(
-                    IdentifierTypes::PAGE_MODEL_ELEMENT_REFERENCE,
-                    new ObjectValue(
-                        ValueTypes::PAGE_ELEMENT_REFERENCE,
-                        'page_model.elements.element_name',
-                        'page_model',
-                        'element_name'
-                    ),
-                    1
-                ),
-                'expectedString' => 'page_model.elements.element_name',
-            ],
-            'page model element reference, position 2' => [
-                'identifier' => new Identifier(
-                    IdentifierTypes::PAGE_MODEL_ELEMENT_REFERENCE,
-                    new ObjectValue(
-                        ValueTypes::PAGE_ELEMENT_REFERENCE,
-                        'page_model.elements.element_name',
-                        'page_model',
-                        'element_name'
-                    ),
-                    2
-                ),
-                'expectedString' => 'page_model.elements.element_name:2',
-            ],
-            'element parameter, position null' => [
-                'identifier' => new Identifier(
-                    IdentifierTypes::ELEMENT_PARAMETER,
-                    new ObjectValue(
-                        ValueTypes::ELEMENT_PARAMETER,
-                        '$elements.element_name',
-                        'elements',
-                        'element_name'
-                    )
-                ),
-                'expectedString' => '$elements.element_name',
-            ],
-            'element parameter, position 1' => [
-                'identifier' => new Identifier(
-                    IdentifierTypes::ELEMENT_PARAMETER,
-                    new ObjectValue(
-                        ValueTypes::ELEMENT_PARAMETER,
-                        '$elements.element_name',
-                        'elements',
-                        'element_name'
-                    ),
-                    1
-                ),
-                'expectedString' => '$elements.element_name',
-            ],
-            'element parameter, position 2' => [
-                'identifier' => new Identifier(
-                    IdentifierTypes::ELEMENT_PARAMETER,
-                    new ObjectValue(
-                        ValueTypes::ELEMENT_PARAMETER,
-                        '$elements.element_name',
-                        'elements',
-                        'element_name'
-                    ),
-                    2
-                ),
-                'expectedString' => '$elements.element_name:2',
-            ],
             'css selector with element reference, position null' => [
                 'identifier' => $cssSelectorWithElementReference,
                 'expectedString' => '"{{ parent_identifier_name }} .selector"',
@@ -278,86 +195,6 @@ class IdentifierTest extends \PHPUnit\Framework\TestCase
                     null,
                     'new identifier name'
                 ),
-            ],
-        ];
-    }
-
-    /**
-     * @dataProvider isActionableDataProvider
-     */
-    public function testIsActionable(IdentifierInterface $identifier, bool $expectedIsActionable)
-    {
-        $this->assertSame($expectedIsActionable, $identifier->isActionable());
-    }
-
-    public function isActionableDataProvider(): array
-    {
-        $value = new LiteralValue('');
-
-        return [
-            'css selector is actionable' => [
-                'identifier' => new Identifier(IdentifierTypes::CSS_SELECTOR, $value),
-                'expectedIsActionable' => true,
-            ],
-            'xpath expression is actionable' => [
-                'identifier' => new Identifier(IdentifierTypes::XPATH_EXPRESSION, $value),
-                'expectedIsActionable' => true,
-            ],
-            'page model element reference is not actionable' => [
-                'identifier' => new Identifier(IdentifierTypes::PAGE_MODEL_ELEMENT_REFERENCE, $value),
-                'expectedIsActionable' => false,
-            ],
-            'element parameter is actionable' => [
-                'identifier' => new Identifier(IdentifierTypes::ELEMENT_PARAMETER, $value),
-                'expectedIsActionable' => true,
-            ],
-            'page object parameter is not actionable' => [
-                'identifier' => new Identifier(IdentifierTypes::PAGE_OBJECT_PARAMETER, $value),
-                'expectedIsActionable' => false,
-            ],
-            'browser object parameter is not actionable' => [
-                'identifier' => new Identifier(IdentifierTypes::BROWSER_OBJECT_PARAMETER, $value),
-                'expectedIsActionable' => false,
-            ],
-        ];
-    }
-
-    /**
-     * @dataProvider isAssertableDataProvider
-     */
-    public function testIsAssertable(IdentifierInterface $identifier, bool $expectedIsActionable)
-    {
-        $this->assertSame($expectedIsActionable, $identifier->isAssertable());
-    }
-
-    public function isAssertableDataProvider(): array
-    {
-        $value = new LiteralValue('');
-
-        return [
-            'css selector is assertable' => [
-                'identifier' => new Identifier(IdentifierTypes::CSS_SELECTOR, $value),
-                'expectedIsAssertable' => true,
-            ],
-            'xpath expression is assertable' => [
-                'identifier' => new Identifier(IdentifierTypes::XPATH_EXPRESSION, $value),
-                'expectedIsAssertable' => true,
-            ],
-            'page model element reference is not assertable' => [
-                'identifier' => new Identifier(IdentifierTypes::PAGE_MODEL_ELEMENT_REFERENCE, $value),
-                'expectedIsAssertable' => false,
-            ],
-            'element parameter is not assertable' => [
-                'identifier' => new Identifier(IdentifierTypes::ELEMENT_PARAMETER, $value),
-                'expectedIsAssertable' => true,
-            ],
-            'page object parameter is assertable' => [
-                'identifier' => new Identifier(IdentifierTypes::PAGE_OBJECT_PARAMETER, $value),
-                'expectedIsAssertable' => true,
-            ],
-            'browser object parameter is assertable' => [
-                'identifier' => new Identifier(IdentifierTypes::BROWSER_OBJECT_PARAMETER, $value),
-                'expectedIsAssertable' => true,
             ],
         ];
     }
